extended transform class more, organized file structure
