+++ index.html
@@ -10,6 +10,6 @@
     <div>
       <canvas id="canvas-main" width="1200" height="600"></canvas>
     </div>
-    <script type="module" src="/src/index.ts"></script>
+    <script type="module" src="/src/app/index.ts"></script>
   </body>
 </html>
