+++ src/graphics/shader-utils/shader.ts
@@ -1,0 +1,13 @@
+﻿export type Shader = {
+
+    vertexCode: string;
+    fragmentCode: string
+
+    vertexModule: GPUShaderModule;
+    fragmentModule: GPUShaderModule
+
+    vertexState: GPUVertexState;
+    fragmentState: GPUFragmentState;
+
+}
+
