+++ src/scene/IRenderable.ts
@@ -1,4 +1,4 @@
-import type {Mesh} from "./mesh.ts";
+import type {Mesh} from "../graphics/3d/mesh.ts";
 import type {Mat4} from "wgpu-matrix";
 
 export interface IRenderable {
